refactor: Move some test helpers to CustomMatchersForSorting.ts
