+++ tests/Sort.test.ts
@@ -7,15 +7,18 @@ window.moment = moment;
 
 import type { Comparator } from '../src/Query/Sorting';
 import { Sort } from '../src/Query/Sort';
-import { DateField } from '../src/Query/Filter/DateField';
 import { Sorting } from '../src/Query/Sorting';
 import { resetSettings, updateSettings } from '../src/Config/Settings';
-import { DateParser } from '../src/Query/DateParser';
 import type { Task } from '../src/Task';
 import { StatusField } from '../src/Query/Filter/StatusField';
 import { DueDateField } from '../src/Query/Filter/DueDateField';
 import { fromLine } from './TestHelpers';
 import { TaskBuilder } from './TestingTools/TaskBuilder';
+import {
+    testDateComparesAfter,
+    testDateComparesBefore,
+    testDateComparesEqual,
+} from './CustomMatchers/CustomMatchersForSorting';
 
 describe('Sort', () => {
     it('constructs Sorting both ways from Comparator function', () => {
@@ -235,44 +238,10 @@ describe('Sort', () => {
     });
 });
 
-expect.extend({
-    toGiveCompareToResult(dates: (string | null)[], expected: -1 | 0 | 1) {
-        expect(dates.length).toEqual(2);
-
-        const dateA = dates[0];
-        const dateB = dates[1];
-
-        let a: moment.Moment | null = null;
-        if (dateA !== null) a = DateParser.parseDate(dateA);
-
-        let b: moment.Moment | null = null;
-        if (dateB !== null) b = DateParser.parseDate(dateB);
-
-        const actual = DateField.compareByDate(a, b);
-
-        const pass = actual === expected;
-        const message = () => `${dateA} < ${dateB}: expected=${expected} actual=${actual}`;
-
-        return { pass, message };
-    },
-});
-
-declare global {
-    namespace jest {
-        interface Matchers<R> {
-            toGiveCompareToResult(expected: number): R;
-        }
-    }
-}
-
 // These are lower-level tests that the Task-based ones above, for ease of test coverage.
 // TODO Replace this with something simpler but equivalent in the tests for DateField.test.ts.
 describe('compareBy', () => {
     it('compares correctly by date', () => {
-        const equal = 0;
-        const after = 1;
-        const before = -1;
-
         const earlierDate = '2022-01-01';
         const latererDate = '2022-02-01'; // intentional typo - laterer - so all variable names align in code
         const invalidDate = '2022-02-30';
@@ -287,25 +256,6 @@ describe('compareBy', () => {
         testDateComparesBefore(invalidDate, null); // invalid dates sort before no date
         testDateComparesEqual(invalidDate, invalidDate);
         testDateComparesAfter(invalidDate, earlierDate); // invalid dates sort after valid ones
-
-        function testDateComparesBefore(dateA: string | null, dateB: string | null) {
-            testCompareByDateBothWays(dateA, dateB, before);
-        }
-
-        function testDateComparesEqual(dateA: string | null, dateB: string | null) {
-            testCompareByDateBothWays(dateA, dateB, equal);
-        }
-
-        function testDateComparesAfter(dateA: string | null, dateB: string | null) {
-            testCompareByDateBothWays(dateA, dateB, after);
-        }
-
-        function testCompareByDateBothWays(dateA: string | null, dateB: string | null, expected: -1 | 0 | 1) {
-            expect([dateA, dateB]).toGiveCompareToResult(expected);
-
-            const reverseExpected = expected === equal ? equal : -expected;
-            expect([dateB, dateA]).toGiveCompareToResult(reverseExpected);
-        }
     });
 });
